+++ CHANGELOG.md
@@ -105,6 +105,8 @@
     [#2429](https://github.com/Kong/kong-operator/pull/2429)
   - Implement adoption support for `KongCertificate`, `KongCACertificate` and `KongSNI`
     [#2484](https://github.com/Kong/kong-operator/pull/2484)
+  - Implement adoption support for `KongVault`.
+    [#2490](https://github.com/Kong/kong-operator/pull/2490)
 - HybridGateway:
   - Added controller-runtime watches for Gateway and GatewayClass resources to the hybridgateway controller.
   - HTTPRoutes are now reconciled when related Gateway or GatewayClass resources change.